requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-jose
requests
pytest
pytest-asyncio
pytest-cov
python-multipart
passlib
bcrypt
pyjwt